Swedish translation for Sfen. Some testing done.
Did not bother with the crowdin site, a direct pull request (why cant I upload the file?)
Used Google Translate, then fixed file, then reviewed in Android Studio 0.8.7. Will be easier when the editor allows in-row editing

Not too happy with translation for Event, Action but no better came up.

Added no-translate for a few strings.xml
Some strings do not look too nice due to casing or that options cannot be inserted the same way in Swedish, that is later